refactor tree view state to be private static in tree view renderer and not in parent
