+++ libs/error-analysis/src/lib/ErrorAnalysisDashboard/Controls/ErrorAnalysisView/ErrorAnalysisView.tsx
@@ -17,7 +17,6 @@ import React from "react";
 import { ErrorAnalysisOptions } from "../../ErrorAnalysisEnums";
 import { HelpMessageDict } from "../../Interfaces/IStringsParam";
 import { IMatrixAreaState, IMatrixFilterState } from "../../MatrixFilterState";
-import { ITreeViewRendererState } from "../../TreeViewState";
 import { MatrixFilter } from "../Matrix/MatrixFilter/MatrixFilter";
 import { TreeViewRenderer } from "../TreeViewRenderer/TreeViewRenderer";
 
@@ -47,8 +46,6 @@ export interface IErrorAnalysisViewProps {
   ) => void;
   selectedCohort: ErrorCohort;
   baseCohort: ErrorCohort;
-  treeViewState: ITreeViewRendererState;
-  setTreeViewState: (treeViewState: ITreeViewRendererState) => void;
   matrixFilterState: IMatrixFilterState;
   matrixAreaState: IMatrixAreaState;
   setMatrixAreaState: (matrixAreaState: IMatrixAreaState) => void;
@@ -76,8 +73,6 @@ export class ErrorAnalysisView extends React.Component<IErrorAnalysisViewProps> 
             updateSelectedCohort={this.props.updateSelectedCohort}
             selectedCohort={this.props.selectedCohort}
             baseCohort={this.props.baseCohort}
-            state={this.props.treeViewState}
-            setTreeViewState={this.props.setTreeViewState}
             showCohortName={this.props.showCohortName}
             disabledView={this.props.disabledView}
           />
@@ -107,7 +102,6 @@ export class ErrorAnalysisView extends React.Component<IErrorAnalysisViewProps> 
   public componentDidUpdate(prevProps: IErrorAnalysisViewProps): void {
     if (
       this.props.selectedFeatures !== prevProps.selectedFeatures ||
-      this.props.treeViewState !== prevProps.treeViewState ||
       this.props.matrixFilterState !== prevProps.matrixFilterState ||
       this.props.matrixAreaState !== prevProps.matrixAreaState
     ) {
