+++ libs/error-analysis/src/lib/ErrorAnalysisDashboard/Controls/ErrorAnalysisView/ErrorAnalysisViewTab.tsx
@@ -118,8 +118,6 @@ export class ErrorAnalysisViewTab extends React.Component<
             updateSelectedCohort={this.props.updateSelectedCohort}
             selectedCohort={this.props.selectedCohort}
             baseCohort={this.props.baseCohort}
-            treeViewState={this.props.treeViewState}
-            setTreeViewState={this.props.setTreeViewState}
             matrixFilterState={this.props.matrixFilterState}
             matrixAreaState={this.props.matrixAreaState}
             setMatrixAreaState={this.props.setMatrixAreaState}
