+++ libs/error-analysis/src/lib/ErrorAnalysisDashboard/Controls/TreeViewRenderer/TreeViewRenderer.tsx
@@ -2,7 +2,6 @@
 // Licensed under the MIT License.
 
 import {
-  ICompositeFilter,
   IFilter,
   FilterMethods,
   CohortSource,
@@ -30,7 +29,6 @@ import { linkVertical as d3linkVertical } from "d3-shape";
 import {
   getTheme,
   IProcessedStyleSet,
-  ITheme,
   mergeStyles,
   MessageBar,
   MessageBarType,
@@ -41,50 +39,26 @@ import React from "react";
 
 import { ColorPalette } from "../../ColorPalette";
 import { FilterProps } from "../../FilterProps";
-import { HelpMessageDict } from "../../Interfaces/IStringsParam";
-import {
-  INodeDetail,
-  ITreeNode,
-  ITreeViewRendererState
-} from "../../TreeViewState";
 import { TreeLegend } from "../TreeLegend/TreeLegend";
 
 import { TreeViewNode } from "./TreeViewNode";
+import { ITreeViewRendererProps } from "./TreeViewProps";
 import {
   ITreeViewRendererStyles,
   treeViewRendererStyles
 } from "./TreeViewRenderer.styles";
+import {
+  createInitialTreeViewState,
+  INodeDetail,
+  ITreeNode,
+  ITreeViewRendererState
+} from "./TreeViewState";
 
 // Importing this solely to set the selectedPanelId. This component is NOT a statefulContainer
 // import StatefulContainer from '../../ap/mixins/statefulContainer.js'
 
 const viewerHeight = 300;
 const viewerWidth = 800;
-
-export interface ITreeViewRendererProps {
-  disabledView: boolean;
-  theme?: ITheme;
-  messages?: HelpMessageDict;
-  features: string[];
-  selectedFeatures: string[];
-  getTreeNodes?: (
-    request: any[],
-    abortSignal: AbortSignal
-  ) => Promise<IErrorAnalysisTreeNode[]>;
-  tree?: IErrorAnalysisTreeNode[];
-  updateSelectedCohort: (
-    filters: IFilter[],
-    compositeFilters: ICompositeFilter[],
-    source: CohortSource,
-    cells: number,
-    cohortStats: MetricCohortStats | undefined
-  ) => void;
-  selectedCohort: ErrorCohort;
-  baseCohort: ErrorCohort;
-  state: ITreeViewRendererState;
-  setTreeViewState: (treeViewState: ITreeViewRendererState) => void;
-  showCohortName: boolean;
-}
 
 export interface ISVGDatum {
   width: number;
@@ -102,12 +76,19 @@ export class TreeViewRenderer extends React.PureComponent<
   ITreeViewRendererState
 > {
   public static contextType = ModelAssessmentContext;
+  private static savedState: ITreeViewRendererState | undefined;
   public context: React.ContextType<typeof ModelAssessmentContext> =
     defaultModelAssessmentContext;
   public constructor(props: ITreeViewRendererProps) {
     super(props);
-    // Note: we take state from props in case
-    this.state = this.props.state;
+    if (
+      this.props.selectedCohort !== this.props.baseCohort &&
+      TreeViewRenderer.savedState
+    ) {
+      this.state = TreeViewRenderer.savedState;
+    } else {
+      this.state = createInitialTreeViewState(this.context.errorAnalysisData);
+    }
   }
 
   public componentDidMount(): void {
@@ -124,12 +105,12 @@ export class TreeViewRenderer extends React.PureComponent<
     if (
       this.props.selectedFeatures !== prevProps.selectedFeatures ||
       this.props.baseCohort !== prevProps.baseCohort ||
-      this.props.state !== prevProps.state ||
-      this.context.errorAnalysisData!.maxDepth !== this.state.maxDepth ||
-      this.context.errorAnalysisData!.numLeaves !== this.state.numLeaves ||
-      this.context.errorAnalysisData!.minChildSamples !==
-        this.state.minChildSamples ||
-      this.context.errorAnalysisData!.metric !== this.state.metric
+      (this.context.errorAnalysisData &&
+        (this.context.errorAnalysisData.maxDepth !== this.state.maxDepth ||
+          this.context.errorAnalysisData.numLeaves !== this.state.numLeaves ||
+          this.context.errorAnalysisData.minChildSamples !==
+            this.state.minChildSamples ||
+          this.context.errorAnalysisData.metric !== this.state.metric))
     ) {
       this.fetchTreeNodes();
     }
@@ -137,7 +118,7 @@ export class TreeViewRenderer extends React.PureComponent<
 
   public componentWillUnmount(): void {
     window.removeEventListener("resize", this.onResize);
-    this.props.setTreeViewState(this.state);
+    TreeViewRenderer.savedState = this.state;
   }
 
   public render(): React.ReactNode {
